Create It_Begins.py
Here we go. Trying to create an AI that will allow for nuance explanation and suggest better words to use while communicating to someone that is not extremely fluent in your language. Translations will occur. I am trying to use the power of Quantum computing to assist in the recognition and fast comprehension of different languages.